requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-rest-auth==2.1.9
Django==4.2
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
Pillow==8.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==2.0.0
setuptools==71.1.0
sqlparse==0.5.1
